Restructure application routing (#1141)
Sets the groundwork for this issue:
https://github.com/lucyparsons/OpenOversight/issues/385

Simplifying the import logic for the `auth` and `main` routes of the
application. This also sets up the work for this PR:
https://github.com/lucyparsons/OpenOversight/pull/1138

- [x] This branch is up-to-date with the `develop` branch.
- [x] `pytest` passes on my local development environment.
- [x] `pre-commit` passes on my local development environment.
