+++ OpenOversight/app/__init__.py
@@ -6,19 +6,19 @@ from logging.handlers import RotatingFileHandler
 from flask import Flask, jsonify, render_template, request
 from flask_bootstrap import Bootstrap5
 from flask_compress import Compress
-from flask_limiter import Limiter
-from flask_limiter.util import get_remote_address
 from flask_login import LoginManager
 from flask_migrate import Migrate
-from flask_sitemap import Sitemap
 from flask_wtf.csrf import CSRFProtect
 
+from OpenOversight.app.auth.views import auth as auth_blueprint
 from OpenOversight.app.email_client import EmailClient
 from OpenOversight.app.filters import instantiate_filters
+from OpenOversight.app.main.views import main as main_blueprint
 from OpenOversight.app.models.config import config
 from OpenOversight.app.models.database import db
 from OpenOversight.app.models.users import AnonymousUser
 from OpenOversight.app.utils.constants import MEGABYTE
+from OpenOversight.app.utils.flask import limiter, sitemap
 
 
 bootstrap = Bootstrap5()
@@ -29,11 +29,6 @@ login_manager.session_protection = "strong"
 login_manager.anonymous_user = AnonymousUser
 login_manager.login_view = "auth.login"
 
-limiter = Limiter(
-    key_func=get_remote_address, default_limits=["100 per minute", "5 per second"]
-)
-
-sitemap = Sitemap()
 csrf = CSRFProtect()
 
 
@@ -52,15 +47,11 @@ def create_app(config_name="default"):
     sitemap.init_app(app)
     compress.init_app(app)
 
-    from OpenOversight.app.main import main as main_blueprint
-
+    # Register Blueprints for application routes
+    app.register_blueprint(auth_blueprint)
     app.register_blueprint(main_blueprint)
 
-    from OpenOversight.app.auth import auth as auth_blueprint
-
-    app.register_blueprint(auth_blueprint, url_prefix="/auth")
-
-    max_log_size = 10 * MEGABYTE  # start new log file after 10 MB
+    max_log_size = 10 * MEGABYTE  # Start new log file after 10 MB
     num_logs_to_keep = 5
     file_handler = RotatingFileHandler(
         "/tmp/openoversight.log", "a", max_log_size, num_logs_to_keep
@@ -77,7 +68,7 @@ def create_app(config_name="default"):
     app.logger.addHandler(file_handler)
     app.logger.info("OpenOversight startup")
 
-    # Also log when endpoints are getting hit hard
+    # Log when endpoints are getting hit hard
     limiter.logger.addHandler(file_handler)
 
     # Define error handlers
