--- OpenOversight/app/auth/__init__.py
@@ -1,6 +1,0 @@
-from flask import Blueprint
-
-
-auth = Blueprint("auth", __name__)
-
-from . import views  # noqa: F401, E402
